+++ backend/tests/api/test_booking.py
@@ -12,12 +12,13 @@ from lm_backend.storage import database
 
 @mark.asyncio
 @database.transaction(force_rollback=True)
-async def test_get_bookings_job(
+async def test_get_bookings_job__success(
     backend_client: AsyncClient,
     some_licenses,
     some_booking_rows,
     some_config_rows,
     insert_objects,
+    inject_security_header,
 ):
     """
     Do I fetch a booking?
@@ -25,7 +26,10 @@ async def test_get_bookings_job(
     await insert_objects(some_licenses, table_schemas.license_table)
     await insert_objects(some_config_rows, table_schemas.config_table)
     await insert_objects(some_booking_rows, table_schemas.booking_table)
+
+    inject_security_header("owner1", "license-manager:booking:read")
     resp = await backend_client.get("/lm/api/v1/booking/job/coolbeans")
+
     assert resp.status_code == 200
     assert resp.json() == [
         dict(
@@ -43,12 +47,40 @@ async def test_get_bookings_job(
 
 @mark.asyncio
 @database.transaction(force_rollback=True)
-async def test_get_bookings_for_cluster_name(
+async def test_get_bookings_job__fails_with_bad_permission(
     backend_client: AsyncClient,
     some_licenses,
     some_booking_rows,
     some_config_rows,
     insert_objects,
+    inject_security_header,
+):
+    """
+    Do I give a 401 with an invalid or missing permission?
+    """
+    await insert_objects(some_licenses, table_schemas.license_table)
+    await insert_objects(some_config_rows, table_schemas.config_table)
+    await insert_objects(some_booking_rows, table_schemas.booking_table)
+
+    # No permission
+    resp = await backend_client.get("/lm/api/v1/booking/job/coolbeans")
+    assert resp.status_code == 401
+
+    # Bad permission
+    inject_security_header("owner1", "invalid-permission")
+    resp = await backend_client.get("/lm/api/v1/booking/job/coolbeans")
+    assert resp.status_code == 403
+
+
+@mark.asyncio
+@database.transaction(force_rollback=True)
+async def test_get_bookings_for_cluster_name__success(
+    backend_client: AsyncClient,
+    some_licenses,
+    some_booking_rows,
+    some_config_rows,
+    insert_objects,
+    inject_security_header,
 ):
     """
     Do I fetch a booking using the cluster_name?
@@ -67,7 +99,10 @@ async def test_get_bookings_for_cluster_name(
         config_id=2,
     )
     await insert_objects([booking], table_schemas.booking_table)
+
+    inject_security_header("owner1", "license-manager:booking:read")
     resp = await backend_client.get("/lm/api/v1/booking/all?cluster_name=cluster2")
+
     assert resp.status_code == 200
     assert resp.json() == [
         dict(
@@ -81,6 +116,44 @@ async def test_get_bookings_for_cluster_name(
             cluster_name="cluster2",
         ),
     ]
+
+
+@mark.asyncio
+@database.transaction(force_rollback=True)
+async def test_get_bookings__fails_on_bad_permission(
+    backend_client: AsyncClient,
+    some_licenses,
+    some_booking_rows,
+    some_config_rows,
+    insert_objects,
+    inject_security_header,
+):
+    """
+    Do I fail with bad or missing permissions?
+    """
+    await insert_objects(some_licenses, table_schemas.license_table)
+    await insert_objects(some_config_rows, table_schemas.config_table)
+    await insert_objects(some_booking_rows, table_schemas.booking_table)
+    booking = BookingRow(
+        id=4,
+        job_id="99",
+        product_feature="hello.world",
+        booked=1,
+        lead_host="host10",
+        user_name="user10",
+        cluster_name="cluster2",
+        config_id=2,
+    )
+    await insert_objects([booking], table_schemas.booking_table)
+
+    # No Permission
+    resp = await backend_client.get("/lm/api/v1/booking/all?cluster_name=cluster2")
+    assert resp.status_code == 401
+
+    # Bad Permission
+    inject_security_header("owner1", "invalid-permission")
+    resp = await backend_client.get("/lm/api/v1/booking/all?cluster_name=cluster2")
+    assert resp.status_code == 403
 
 
 @mark.asyncio
@@ -103,6 +176,7 @@ async def test_bookings_all(
     some_booking_rows,
     some_config_rows,
     insert_objects,
+    inject_security_header,
 ):
     """
     Do I fetch all the bookings in the db?
@@ -110,7 +184,10 @@ async def test_bookings_all(
     await insert_objects(some_licenses, table_schemas.license_table)
     await insert_objects(some_config_rows, table_schemas.config_table)
     await insert_objects(some_booking_rows, table_schemas.booking_table)
+
+    inject_security_header("owner1", "license-manager:booking:read")
     resp = await backend_client.get("/lm/api/v1/booking/all")
+
     assert resp.status_code == 200
     assert resp.json() == [
         dict(
@@ -148,7 +225,13 @@ async def test_bookings_all(
 
 @mark.asyncio
 @database.transaction(force_rollback=True)
-async def test_booking_create(backend_client, some_config_rows, some_licenses, insert_objects):
+async def test_booking_create__success(
+    backend_client,
+    some_config_rows,
+    some_licenses,
+    insert_objects,
+    inject_security_header,
+):
     """This test proves that a booking can be created by showing that the response status is 200."""
     await insert_objects(some_licenses, table_schemas.license_table)
     await insert_objects(some_config_rows, table_schemas.config_table)
@@ -156,6 +239,8 @@ async def test_booking_create(backend_client, some_config_rows, some_licenses, i
     booking = Booking(
         job_id=1, features=[features], lead_host="host1", user_name="user1", cluster_name="cluster1"
     )
+
+    inject_security_header("owner1", "license-manager:booking:write")
     resp = await backend_client.put("/lm/api/v1/booking/book", json=booking.dict())
 
     assert resp.status_code == status.HTTP_200_OK
@@ -163,8 +248,42 @@ async def test_booking_create(backend_client, some_config_rows, some_licenses, i
 
 @mark.asyncio
 @database.transaction(force_rollback=True)
+async def test_booking_create__fail_on_bad_permission(
+    backend_client,
+    some_config_rows,
+    some_licenses,
+    insert_objects,
+    inject_security_header,
+):
+    """
+    This test proves that a 401 or 403 will be returend by the endpoint if the supplied
+    auth token is missiong or invalid.
+    """
+    await insert_objects(some_licenses, table_schemas.license_table)
+    await insert_objects(some_config_rows, table_schemas.config_table)
+    features = BookingFeature(booked=10, product_feature="hello.world")
+    booking = Booking(
+        job_id=1, features=[features], lead_host="host1", user_name="user1", cluster_name="cluster1"
+    )
+
+    # No Permission
+    resp = await backend_client.put("/lm/api/v1/booking/book", json=booking.dict())
+    assert resp.status_code == status.HTTP_401_UNAUTHORIZED
+
+    # Bad Permission
+    inject_security_header("owner1", "bad-permission")
+    resp = await backend_client.put("/lm/api/v1/booking/book", json=booking.dict())
+    assert resp.status_code == status.HTTP_403_FORBIDDEN
+
+
+@mark.asyncio
+@database.transaction(force_rollback=True)
 async def test_booking_create_negative_booked_error(
-    backend_client, some_config_rows, some_licenses, insert_objects
+    backend_client,
+    some_config_rows,
+    some_licenses,
+    insert_objects,
+    inject_security_header,
 ):
     """This test proves that a 400 response code is returned when a `-` (negative) booking creation is
     attempted by checking the response code of a booking request containing a negative booking.
@@ -175,6 +294,8 @@ async def test_booking_create_negative_booked_error(
     booking = Booking(
         job_id=1, features=[features], lead_host="host1", user_name="user1", cluster_name="cluster1"
     )
+
+    inject_security_header("owner1", "license-manager:booking:write")
     resp = await backend_client.put("/lm/api/v1/booking/book", json=booking.dict())
 
     assert resp.status_code == status.HTTP_400_BAD_REQUEST
@@ -187,6 +308,7 @@ async def test_booking_create_booked_greater_than_total(
     some_config_rows,
     some_licenses,
     insert_objects,
+    inject_security_header,
 ):
     """This test proves that the correct response (400) is returned when a booking
     request exceeds the total available by asserting that the response detail contains
@@ -198,6 +320,7 @@ async def test_booking_create_booked_greater_than_total(
     booking = Booking(
         job_id=1, features=[features], lead_host="host1", user_name="user1", cluster_name="cluster1"
     )
+    inject_security_header("owner1", "license-manager:booking:write")
     resp = await backend_client.put("/lm/api/v1/booking/book", json=booking.dict())
 
     assert resp.status_code == status.HTTP_400_BAD_REQUEST
@@ -206,12 +329,13 @@ async def test_booking_create_booked_greater_than_total(
 
 @mark.asyncio
 @database.transaction(force_rollback=True)
-async def test_booking_delete(
+async def test_booking_delete__success(
     backend_client,
     some_config_rows,
     some_licenses,
     some_booking_rows,
     insert_objects,
+    inject_security_header,
 ):
     """This test proves that the correct response is returned (200) when a booking
     is successfully deleted.
@@ -220,8 +344,37 @@ async def test_booking_delete(
     await insert_objects(some_config_rows, table_schemas.config_table)
     await insert_objects(some_booking_rows, table_schemas.booking_table)
 
+    inject_security_header("owner1", "license-manager:booking:write")
     resp = await backend_client.delete("/lm/api/v1/booking/book/helloworld")
     assert resp.status_code == status.HTTP_200_OK
+
+
+@mark.asyncio
+@database.transaction(force_rollback=True)
+async def test_booking_delete__fail_on_bad_permissions(
+    backend_client,
+    some_config_rows,
+    some_licenses,
+    some_booking_rows,
+    insert_objects,
+    inject_security_header,
+):
+    """
+    This test proves that a 401 or 403 response is returned when the auth token
+    is missing or invalid.
+    """
+    await insert_objects(some_licenses, table_schemas.license_table)
+    await insert_objects(some_config_rows, table_schemas.config_table)
+    await insert_objects(some_booking_rows, table_schemas.booking_table)
+
+    # No Permission
+    resp = await backend_client.delete("/lm/api/v1/booking/book/helloworld")
+    assert resp.status_code == status.HTTP_401_UNAUTHORIZED
+
+    # Bad Permission
+    inject_security_header("owner1", "invalid-permission")
+    resp = await backend_client.delete("/lm/api/v1/booking/book/helloworld")
+    assert resp.status_code == status.HTTP_403_FORBIDDEN
 
 
 @mark.asyncio